Al eliminar un usuario responde con json

+++ script_eliminar_usuarios.py
@@ -1,5 +1,6 @@
 import argparse
 import pymysql
+import json
 from zk import ZK
 
 # Argumentos
@@ -16,19 +17,34 @@ conexion = pymysql.connect(
     database='pruebaChecador'
 )
 cursor = conexion.cursor()
-print("Conexión exitosa a la base de datos.")
 
 # Conexión al checador
 zk = ZK(args.ip, port=4370, timeout=5, password=0, force_udp=False, ommit_ping=False)
 conn = None
 
 try:
+    response = {}
     conn = zk.connect()
     conn.disable_device()
 
+    # Verificar si el usuario existe en el checador
+    usuarios_dispositivo = conn.get_users()
+    usuario_existente = next((u for u in usuarios_dispositivo if int(u.user_id) == args.id), None)
+    if not usuario_existente:
+        raise Exception(f"El usuario con ID {args.id} no existe en el dispositivo.")
+
     # Eliminar usuario del checador
     conn.delete_user(user_id=args.id)
-    print(f"Usuario con ID {args.id} eliminado del checador.")
+    response["dispositivo"] = {
+        "status": "eliminado",
+        "id": args.id
+    }
+
+    # Verificar si el usuario existe en la base de datos
+    cursor.execute("SELECT COUNT(*) FROM empleados WHERE id_usuario = %s", (args.id,))
+    resultado = cursor.fetchone()
+    if resultado[0] == 0:
+        raise Exception(f"El usuario con ID {args.id} no existe en la base de datos.")
 
     # Eliminar usuario de la base de datos
     cursor.execute("DELETE FROM huellasrespaldo WHERE id_usuario = %s", (args.id,))
@@ -39,15 +55,22 @@ try:
     resultado = cursor.fetchone()
     if resultado[0] == 0:
         conexion.commit()
-        print("Verificación exitosa: el usuario ya no está en la base de datos.")
+        response["base_datos"] = {
+            "status": "eliminado",
+            "id": args.id
+        }
     else:
         conexion.rollback()
-        print("Error: el usuario aún existe en la base de datos. Cambios revertidos.")
+        raise Exception("Error: el usuario aún existe en la base de datos. Cambios revertidos.")
 
-except Exception as e:
-    conexion.rollback()
-    print(f"Proceso terminado con error: {e}")
+    conn.enable_device()
 
+    # Devolver respuesta en formato JSON
+    print(json.dumps(response, indent=4))
+
+except Exception:
+    # Devolver JSON vacío en caso de error
+    print(json.dumps({}))
 finally:
     if conn:
         conn.disconnect()
